Moved transmissibility feature to new subdir.
