+++ opm/grid/transmissibility/TransTpfa_impl.hpp
@@ -1,6 +1,6 @@
 #include <opm/common/utility/numeric/blas_lapack.h>
-#include <opm/core/pressure/tpfa/trans_tpfa.h>
-#include <opm/core/grid/GridHelpers.hpp>
+#include <opm/grid/transmissibility/trans_tpfa.h>
+#include <opm/grid/GridHelpers.hpp>
 
 #include <cmath>
